requirements.txt
# Zależności dla Systemu Rekomendacji Filmów
# Zainstaluj używając: pip install -r requirements.txt

# Komunikacja z API
requests>=2.28.0

# Zmienne środowiskowe
python-dotenv>=0.19.0

# Analiza danych i algorytmy rekomendacji
numpy>=1.21.0

# Opcjonalne zależności dla rozszerzonej funkcjonalności
# scikit-learn>=1.0.0  # Dla zaawansowanych algorytmów ML
# pandas>=1.3.0        # Dla analizy danych 